Changing DAO Interface parameter id to long

+++ library_software/src/main/java/de/frauas/library/data/DAO.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface DAO <T>{
 	
-	Optional<T> get(int id);
+	Optional<T> get(long id);
 	
 	List<T> getAll();
